Publicly expose the shared AFImageCache.
This is a backport (of sorts) of an interface introduced in
421cf892274d7d49d17af7624b80dc24ff70be83 for an application that cannot
yet migrate to AFNetworking >= 2.0, yet needs the ability to
occasionally flush the image cache.

(Another possibility would be to make UIImage conform to
NSDiscardableContent.  That would, in theory, eliminate the need for the
change in this commit, because we could piggyback off of NSCache's
built-in eviction policy.  However, it is not clear to me how UIImage
would be conformed in that way.)